legacy_backend/requirements.txt
Django==5.2.8
Requests==2.32.5
feedparser==6.0.12
playwright==1.56.0
gunicorn==23.0.0
django-environ==0.12.0
beautifulsoup4==4.12.3
whitenoise==6.8.2
praw==7.8.1
Markdown==3.7
django-q2==1.8.0
croniter==5.0.1
ruff==0.14.5
Pillow==11.0.0
pydantic==2.10.6
google-api-python-client==2.154.0
django-ninja==1.3.0
djangoql==0.18.1
